cmake: double-arith titleX/titleY in DrawCmakeTitle (83.4%)

+++ src/cmake.cpp
@@ -3554,8 +3554,8 @@ void CMenuPcs::DrawCmakeTitle(int page, float x, float alpha)
 
     MenuPcs.SetTexture(static_cast<CMenuPcs::TEX>((CmakeResult(this) != 0) ? 0x65 : 0x3E));
 
-    float titleX = static_cast<float>(offsU + static_cast<unsigned int>(20.0));
-    float titleY = static_cast<float>(offsU + static_cast<int>(8.0));
+    float titleX = static_cast<float>(static_cast<int>(static_cast<double>(offsU) + 20.0));
+    float titleY = static_cast<float>(static_cast<int>(static_cast<double>(offsU) + 8.0));
     MenuPcs.DrawRect(
         0, titleX, titleY, 208.0f, 24.0f,
         0.0f, static_cast<float>(page * 0x18), 1.0f, 1.0f, 0.0f);
